* release-1.7: (192 commits)
  Optional value with no default value do not set anything when being mapped.
  Fix warning
  Adds property for versionLabel to be able to customize it in stylesheet.
  Adds setupWithCollection & mappings on viewContainerController to help setuping the controller with a new collection later that init.
  Section extension to create property grid sections without adding it to the form. addSections on form that trigger insertion animations in batch.
  Forms animated section insertion if the view is visible on screen.
  Fix for cell size & flags computing when mixed with stylesheet definitions.
  Fix reserved keywords in styles.
  Adds public method applyStyleForNavigation on CKUIViewCOntroller. This can be called in some cases when the rootViewController is instantaneously hidden by a modal controller. Call this in viewDidLoad instead as navigationController is loaded from nib and viewWillAppear is not called.
  CKWebRequest2: disables the checking of self-signed certificates.
  CKWebRequest2: fixes issue with failure blocks not called when a parsing error occurs; fixes issue with HTTP error 400 not recognized as an error.
  Adds supportedInterfaceOrientations on CKUIViewController.
  Fix class conversion + set in property.
  Fix native class Type 'Class' conversion from string.
  disable debug checkForBlockCopy if not DEBUG
  fix for UITextInputTraits introspection.
  TextView introspection extension.
  ArrayCollection propagate KVO on Count property event in replace object.
  Overrides KVO methods for TextField to support UIInputTraits protocol as key value complient.
  Adds QuartzCore as import in cloudKit.h
  ...

+++ Classes/CKPageScrollView.m
@@ -128,7 +128,7 @@
 }
 
 - (void)insertView:(UIView *)view atIndex:(NSUInteger)index {
-	view.center = CGPointMake((self.pageWidth / 2) + (self.pageWidth * index), self.bounds.size.height / 2);
+	view.center = CGPointMake((NSInteger)((self.pageWidth / 2.0) + (self.pageWidth * index)),(NSInteger)( self.bounds.size.height / 2.0));
 	[self.scrollView addSubview:view];
 }
